+++ lib/rlm/response_protocol.rb
@@ -5,6 +5,7 @@ require_relative "response_protocol/tags"
 require_relative "response_protocol/json"
 require_relative "response_protocol/xml"
 require_relative "response_protocol/native_json"
+require_relative "response_protocol/baml"
 require_relative "response_protocol/selection"
 
 module RLM